feat: style code blocks
