+++ src/template.json
@@ -92,7 +92,97 @@
   },
   "code_block": {
     "color": "$mantle",
-    "margin": 2
+    "margin": 2,
+    "chroma": {
+      "text": {
+        "color": "$text"
+      },
+      "error": {
+        "color": "$text",
+        "background_color": "$red"
+      },
+      "comment": {
+        "color": "$overlay0"
+      },
+      "comment_preproc": {
+        "color": "$blue"
+      },
+      "keyword": {
+        "color": "$mauve"
+      },
+      "keyword_reserved": {
+        "color": "$mauve"
+      },
+      "keyword_namespace": {
+        "color": "$yellow"
+      },
+      "keyword_type": {
+        "color": "$yellow"
+      },
+      "operator": {
+        "color": "$sky"
+      },
+      "punctuation": {
+        "color": "$overlay2"
+      },
+      "name": {
+        "color": "$lavender"
+      },
+      "name_builtin": {
+        "color": "$peach"
+      },
+      "name_tag": {
+        "color": "$mauve"
+      },
+      "name_attribute": {
+        "color": "$yellow"
+      },
+      "name_class": {
+        "color": "$yellow"
+      },
+      "name_constant": {
+        "color": "$yellow"
+      },
+      "name_decorator": {
+        "color": "$pink"
+      },
+      "name_exception": {},
+      "name_function": {
+        "color": "$blue"
+      },
+      "name_other": {},
+      "literal": {},
+      "literal_number": {
+        "color": "$peach"
+      },
+      "literal_date": {},
+      "literal_string": {
+        "color": "$green"
+      },
+      "literal_string_escape": {
+        "color": "$pink"
+      },
+      "generic_deleted": {
+        "color": "$red"
+      },
+      "generic_emph": {
+        "color": "$text",
+        "italic": true
+      },
+      "generic_inserted": {
+        "color": "$green"
+      },
+      "generic_strong": {
+        "color": "$text",
+        "bold": true
+      },
+      "generic_subheading": {
+        "color": "$sky"
+      },
+      "background": {
+        "background_color": "$mantle"
+      }
+    }
   },
   "table": {
     "center_separator": "┼",
