+++ themes/frappe.json
@@ -92,7 +92,97 @@
   },
   "code_block": {
     "color": "#292c3c",
-    "margin": 2
+    "margin": 2,
+    "chroma": {
+      "text": {
+        "color": "#c6d0f5"
+      },
+      "error": {
+        "color": "#c6d0f5",
+        "background_color": "#e78284"
+      },
+      "comment": {
+        "color": "#737994"
+      },
+      "comment_preproc": {
+        "color": "#8caaee"
+      },
+      "keyword": {
+        "color": "#ca9ee6"
+      },
+      "keyword_reserved": {
+        "color": "#ca9ee6"
+      },
+      "keyword_namespace": {
+        "color": "#e5c890"
+      },
+      "keyword_type": {
+        "color": "#e5c890"
+      },
+      "operator": {
+        "color": "#99d1db"
+      },
+      "punctuation": {
+        "color": "#949cbb"
+      },
+      "name": {
+        "color": "#babbf1"
+      },
+      "name_builtin": {
+        "color": "#ef9f76"
+      },
+      "name_tag": {
+        "color": "#ca9ee6"
+      },
+      "name_attribute": {
+        "color": "#e5c890"
+      },
+      "name_class": {
+        "color": "#e5c890"
+      },
+      "name_constant": {
+        "color": "#e5c890"
+      },
+      "name_decorator": {
+        "color": "#f4b8e4"
+      },
+      "name_exception": {},
+      "name_function": {
+        "color": "#8caaee"
+      },
+      "name_other": {},
+      "literal": {},
+      "literal_number": {
+        "color": "#ef9f76"
+      },
+      "literal_date": {},
+      "literal_string": {
+        "color": "#a6d189"
+      },
+      "literal_string_escape": {
+        "color": "#f4b8e4"
+      },
+      "generic_deleted": {
+        "color": "#e78284"
+      },
+      "generic_emph": {
+        "color": "#c6d0f5",
+        "italic": true
+      },
+      "generic_inserted": {
+        "color": "#a6d189"
+      },
+      "generic_strong": {
+        "color": "#c6d0f5",
+        "bold": true
+      },
+      "generic_subheading": {
+        "color": "#99d1db"
+      },
+      "background": {
+        "background_color": "#292c3c"
+      }
+    }
   },
   "table": {
     "center_separator": "┼",
