+++ themes/latte.json
@@ -92,7 +92,97 @@
   },
   "code_block": {
     "color": "#e6e9ef",
-    "margin": 2
+    "margin": 2,
+    "chroma": {
+      "text": {
+        "color": "#4c4f69"
+      },
+      "error": {
+        "color": "#4c4f69",
+        "background_color": "#d20f39"
+      },
+      "comment": {
+        "color": "#9ca0b0"
+      },
+      "comment_preproc": {
+        "color": "#1e66f5"
+      },
+      "keyword": {
+        "color": "#8839ef"
+      },
+      "keyword_reserved": {
+        "color": "#8839ef"
+      },
+      "keyword_namespace": {
+        "color": "#df8e1d"
+      },
+      "keyword_type": {
+        "color": "#df8e1d"
+      },
+      "operator": {
+        "color": "#04a5e5"
+      },
+      "punctuation": {
+        "color": "#7c7f93"
+      },
+      "name": {
+        "color": "#7287fd"
+      },
+      "name_builtin": {
+        "color": "#fe640b"
+      },
+      "name_tag": {
+        "color": "#8839ef"
+      },
+      "name_attribute": {
+        "color": "#df8e1d"
+      },
+      "name_class": {
+        "color": "#df8e1d"
+      },
+      "name_constant": {
+        "color": "#df8e1d"
+      },
+      "name_decorator": {
+        "color": "#ea76cb"
+      },
+      "name_exception": {},
+      "name_function": {
+        "color": "#1e66f5"
+      },
+      "name_other": {},
+      "literal": {},
+      "literal_number": {
+        "color": "#fe640b"
+      },
+      "literal_date": {},
+      "literal_string": {
+        "color": "#40a02b"
+      },
+      "literal_string_escape": {
+        "color": "#ea76cb"
+      },
+      "generic_deleted": {
+        "color": "#d20f39"
+      },
+      "generic_emph": {
+        "color": "#4c4f69",
+        "italic": true
+      },
+      "generic_inserted": {
+        "color": "#40a02b"
+      },
+      "generic_strong": {
+        "color": "#4c4f69",
+        "bold": true
+      },
+      "generic_subheading": {
+        "color": "#04a5e5"
+      },
+      "background": {
+        "background_color": "#e6e9ef"
+      }
+    }
   },
   "table": {
     "center_separator": "┼",
